Succesful run of publish_report

+++ bus_service_increase/publish_single_report.py
@@ -29,48 +29,52 @@ DEFAULT_COMMITTER = {
 notebooks_to_run = {
     # key: name of notebook in directory
     # value: name of the notebook papermill execution (can be renamed)
-    "highways-no-parallel-routes-gh.ipynb": "./highways-no-parallel-routes.ipynb",
-    "highways-low-competitive-routes.ipynb": "./highways-low-competitive-routes.ipynb"
+    #"highways-no-parallel-routes-gh.ipynb": "highways-no-parallel-routes.ipynb",
+    "highways-low-competitive-routes.ipynb": "highways-low-competitive-routes.ipynb"
 }
 
 def publish_notebooks(notebooks_to_run):
-    try:
-        for input_file, output_file in notebooks_to_run.items():        
-            pm.execute_notebook(
-                input_file,
-                output_file,
-                cwd=f"./{REPORT_FOLDER}/",
-                #log_output=True
-            )
-            output_format = 'html'
+    for input_file, output_file in notebooks_to_run.items():        
+        pm.execute_notebook(
+            input_file,
+            output_file,
+            #cwd=f"./{REPORT_FOLDER}/",
+            #log_output=True
+        )
+        output_format = 'html'
 
-            subprocess.run([
-                "jupyter",
-                "nbconvert",
-                "--to",
-                output_format,
-                "--no-input",
-                "--no-prompt",
-                output_file,
-            ]) 
+        subprocess.run([
+            "jupyter",
+            "nbconvert",
+            "--to",
+            output_format,
+            "--no-input",
+            "--no-prompt",
+            output_file,
+        ]) 
 
-            print("Converted to HTML")
-            # Now find the HTML file and upload
-            html_file_name = output_file.replace(".ipynb", ".html")
-            
-            utils.upload_file_to_github(
-                TOKEN,
-                REPO,
-                BRANCH,
-                f"{REPO_FOLDER}{REPORT_FOLDER}{html_file_name}",
-                f"./{REPORT_ROLDER}{html_file_name}",
-                f"Upload {html_file_name}",
-                DEFAULT_COMMITTER
-            )
-            print("Successful upload to GitHub")
+        print("Converted to HTML")
+        # Now find the HTML file and upload
+        html_file_name = output_file.replace(".ipynb", ".html")
+    
+        # Move the html file into the same folder as where it's stored for GH pages rendering
+        # Can't move before because functions won't import 
+        os.rename(f"{html_file_name}", f"{REPORT_FOLDER}{html_file_name}")
+
+        utils.upload_file_to_github(
+            TOKEN,
+            REPO,
+            BRANCH,
+            f"{REPO_FOLDER}{REPORT_FOLDER}{html_file_name}",
+            f"{REPORT_FOLDER}{html_file_name}",
+            f"Upload {html_file_name}",
+            DEFAULT_COMMITTER
+        )
+        print("Successful upload to GitHub")
+        
+        # Clean up - since I renamed file, I don't want executed notebook to stay in directory
+        if input_file != output_file:
             os.remove(output_file)
-    except:
-        print(f"Unsuccessful upload of {html_file_name}")
     
     
 if __name__ == "__main__":
